Re-enable WebP support
Only keep it disabled for the legacy macOS 10.12 builds since those
use an older Qt5 version.

+++ scripts/package/macos-package-app.sh
@@ -69,9 +69,11 @@ pushd "$APP_BUNDLE/Contents/MacOS/PyQt5/$QT5_DIR/"
 ln -s "../../../Resources/$QT5_DIR/translations" .
 popd
 
-# Mitigate libwebp vulnerability allowing for arbitrary code execution (CVE-2023-4863).
-# Disable the Qt webp imageformat plugin.
-rm "$APP_BUNDLE/Contents/MacOS/PyQt5/$QT5_DIR/plugins/imageformats/libqwebp.dylib"
+if [ "$DISABLE_WEBP" = '1' ]; then
+  # Mitigate libwebp vulnerability allowing for arbitrary code execution (CVE-2023-4863).
+  # Disable the Qt webp imageformat plugin if requested.
+  rm "$APP_BUNDLE/Contents/MacOS/PyQt5/$QT5_DIR/plugins/imageformats/libqwebp.dylib"
+fi
 
 if [ "$CODESIGN" = '1' ]; then
     # Enable hardened runtime if app will get notarized
